pyproject.toml
[tool.poetry]
name = "foqa"
description = "Faroese question-answering dataset."
version = "0.0.0"
authors = [
    "Dan Saattrup Nielsen <dan.nielsen@alexandra.dk>",
]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
hydra-core = "^1.3.2"
openai = "^1.21.2"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
datasets = "^2.18.0"

[build-system]
requires = [
    "poetry-core>=1.0.0"
]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.1.1"
pytest-cov = ">=4.1.0"
pre-commit = ">=3.6.2"
pdoc = ">=14.1.0"
readme-coverage-badger = ">=0.1.2"
click = ">=8.1.7"
ruff = ">=0.3.2"
mypy = ">=1.9.0"
nbstripout = ">=0.7.1"

[[tool.poetry.source]]
name = "pypi"


[tool.ruff]
target-version = "py311"
line-length = 88
exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
skip-magic-trailing-comma = true

[tool.ruff.lint]
extend-select = [
    "I",
    "D",
]

[tool.ruff.lint.extend-per-file-ignores]
"__init__.py" = [
    "F401",
]

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    '--verbose',
    '--durations=10',
    '--color=yes',
    '-s',
    '-vv',
    '--doctest-modules',
    '--cov=src/foqa',
]
xfail_strict = true
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::ImportWarning",
]
log_cli_level = "info"
testpaths = [
    "tests",
    "src/foqa",
]
